requirements.txt
streamlit
ultralytics
opencv-python
pillow
torch
torchvision
pandas
seaborn
scikit-learn